requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-rest-swagger
django-filter==22.1
drf-yasg==1.21.4
Pillow==9.2.0
psycopg2-binary==2.9.3
